move getWeeklyPicks function to authFunctions file, delete crudFunctions directory, revert authButton import change, remove process.env variables and global interface, remove unnecessary properties in getWeeklyPicks json fetch
